requirements.txt
# Requirements for the sentiment analysis code
pytorch
scikit-learn
transformers == 3.*
flask
